+++ src/main/java/com/example/fieldpasserbe/member/service/EmailService.java
@@ -1,4 +1,4 @@
-package com.example.fieldpasserbe.service;
+package com.example.fieldpasserbe.member.service;
 
 public interface EmailService {
     String sendSimpleMessage(String to)throws Exception;
